+++ scripts/ex/processing_metrics/ex_dsc_coverage_metrics.py
@@ -8,12 +8,13 @@ Authors:
     - Joshua Johnstone
     - Chat-GPT
 """
+
 # Import libraries
 import sys
-import bisect
-import json
 import argparse
-import pysam
+import numpy as np
+import subprocess
+import json
 import plotly.graph_objects as go
 
 def main(args):
@@ -23,112 +24,175 @@ def main(args):
     sys.stderr = open(args.log, "a")
     print("[INFO] Starting ex_dsc_coverage_metrics.py")
 
-    # Load quality threshold
-    BASE_QUALITY_THRESHOLD = int(args.base_quality_threshold)
-
-    # Inputs from Snakemake
-    bam_ex_dsc_path = args.bam_ex_dsc
+    # Define input paths  
+    ex_dsc_bam_path = args.ex_dsc_bam
     include_bed_path = args.include_bed
     ref_fai_path = args.ref_fai
-    sample = args.sample
 
-    # Output paths
-    json_out_path = args.json
-    plot_out_path = args.plot
+    # Define output paths
+    json_out_path = args.output_json
+    plot_out_path = args.output_plot
+
+    # Define params
+    EX_DEPTH_THRESHOLD = int(args.ex_depth_threshold)
+    EX_BQ_THRESHOLD = int(args.ex_bq_threshold)
+    THREADS = int(args.threads)
 
     # Helper functions
-    # Loads BED intervals to dict
-    def load_bed_intervals(path):
-        bed_dict = {}
-        with open(path) as f:
+    # Returns a dict with [chrom][length] from FAI file
+    def get_chrom_lengths(fai_path):
+        chrom_lengths = {}
+        with open(fai_path) as f:
             for line in f:
-                # Skip empty lines
-                if not line.strip():
-                    continue
-                # Extract chrom, start, end
-                chrom, start, end = line.split()[:3]
-                start, end = int(start), int(end)
-                bed_dict.setdefault(chrom, []).append((start, end))
-        # Sort by start pos
-        for chrom in bed_dict:
-            bed_dict[chrom].sort()
-        return bed_dict
+                chrom, length = line.strip().split("\t")[:2]
+                chrom_lengths[chrom] = int(length)
+        return chrom_lengths
 
-    # Checks if a position exists in BED intervals
-    def in_intervals(chrom, pos, bed_dict):
-        intervals = bed_dict.get(chrom, [])
-        # Get last interval with start <= pos
-        i = bisect.bisect_right(intervals, (pos, float('inf'))) - 1
-        # Return TRUE if pos lies inside interval
-        return i >= 0 and intervals[i][0] <= pos < intervals[i][1]
-
-    # Get total genome positions from FAI
-    ref_lengths = {}
-    with open(ref_fai_path) as f:
-        for line in f:
-            chrom, length = line.split()[:2]
-            ref_lengths[chrom] = int(length)
-    total_genome_positions = sum(ref_lengths.values())
-
-    # Load include BED intervals
-    include_intervals = load_bed_intervals(include_bed_path)
+    # Returns a dict with [chrom][start_index], and total genome length
+    def get_chrom_offsets(chrom_lengths):
+        offsets = {}
+        genome_length = 0
+        for chrom, length in chrom_lengths.items():
+            offsets[chrom] = genome_length
+            genome_length += length
+        return offsets, genome_length
     
-    # Initialise counters
-    ex_dsc_coverage_bases = 0
-    ex_dsc_high_qual_bases = 0
-    ex_dsc_high_qual_unmasked_bases = 0
+    # Creates a boolean array for BED file coverage
+    def coverage_array_bed(bed_path, chrom_lengths):
 
-    # Stream DSC BAM and check criteria for each position
-    bam_ex_dsc = pysam.AlignmentFile(bam_ex_dsc_path, "rb")
-
-    for position_pileup in bam_ex_dsc.pileup(stepper = "all", truncate = True):
-
-        # Extract position and read pileup for this position
-        chrom = position_pileup.reference_name
-        pos = position_pileup.reference_pos
-        depth = position_pileup.nsegments
-        reads = position_pileup.pileups
-
-        # Check each criterion for this position
-        ex_dsc_coverage_check = False
-        if depth > 0:
-            ex_dsc_coverage_check = True
-
-        ex_dsc_high_qual_check = False
-        if ex_dsc_coverage_check:
-            for read in reads:
-                if not read.is_del and not read.is_refskip:
-                    if read.alignment.query_qualities[read.query_position] >= BASE_QUALITY_THRESHOLD:
-                        ex_dsc_high_qual_check = True
-                        break
+        print(f"[INFO] Started creating coverage array for {bed_path}")
         
-        ex_dsc_high_qual_unmasked_check = False
-        if ex_dsc_high_qual_check and in_intervals(chrom, pos, include_intervals):
-            ex_dsc_high_qual_unmasked_check = True
+        # Get chromosome offsets to caclulate array indices
+        offsets, genome_length = get_chrom_offsets(chrom_lengths)
 
-        # Increment counters if criteria met
-        if ex_dsc_coverage_check:
-            ex_dsc_coverage_bases += 1
+        # Set coverage to False for all positions
+        coverage_array = np.zeros(genome_length, dtype=bool)
 
-        if ex_dsc_high_qual_check:
-            ex_dsc_high_qual_bases += 1
+        # Mark BED-covered positions as True
+        with open(bed_path) as bed:
+            for line in bed:
+                if line.startswith("#") or not line.strip():
+                    continue
 
-        if ex_dsc_high_qual_unmasked_check:
-            ex_dsc_high_qual_unmasked_bases += 1
+                chrom, start, end = line.rstrip().split()[:3]
+                start = int(start)
+                end = int(end)
 
-    bam_ex_dsc.close()
+                genome_start = offsets[chrom] + start
+                genome_end = offsets[chrom] + end
 
-    # Calculate metrics
-    ex_dsc_coverage_pct = round((ex_dsc_coverage_bases / total_genome_positions) * 100, ndigits = 2)
-    ex_dsc_high_qual_pct = round((ex_dsc_high_qual_bases / total_genome_positions) * 100, ndigits = 2)
-    ex_dsc_high_qual_unmasked_pct = round((ex_dsc_high_qual_unmasked_bases / total_genome_positions) * 100, ndigits = 2)
+                coverage_array[genome_start:genome_end] = True
+
+        print(f"[INFO] Finished creating coverage array for {bed_path}")
+
+        return coverage_array
+    
+    # Creates a boolean array for coverage at each BAM position
+    def coverage_array_depth(bam_path, chrom_lengths, depth_threshold, threads):
+
+        print(f"[INFO] Started creating depth coverage array for {bam_path}")
+        
+        # Get chromosome offsets to caclulate array indices
+        offsets, genome_length = get_chrom_offsets(chrom_lengths)
+
+        # Set coverage to False for all positions
+        coverage_array_depth = np.zeros(genome_length, dtype=bool)
+
+        cmd = [
+        "samtools", "depth",
+        "--threads", str(threads),
+        "-J",
+        "-s",
+        bam_path
+        ]
+
+        with subprocess.Popen(
+            cmd,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+            text=True
+        ) as proc:
+            for line in proc.stdout:
+                chrom, pos_str, depth_str = line.split()
+                pos = int(pos_str) - 1 # Convert position to 0-based
+                depth = int(depth_str)
+                genome_index = offsets[chrom] + pos
+
+                # If depth >= threshold, set coverage to True
+                if depth >= depth_threshold:
+                    coverage_array_depth[genome_index] = True
+
+        print(f"[INFO] Finished creating depth coverage array for {bam_path}")
+
+        return coverage_array_depth
+    
+    # Creates a boolean array for coverage at each BAM position (at a given BQ threshold)
+    def coverage_array_BQ_threshold(bam_path, chrom_lengths, depth_threshold, BQ_threshold, threads):
+
+        print(f"[INFO] Started creating BQ coverage array for {bam_path}")
+        
+        # Get chromosome offsets to caclulate array indices
+        offsets, genome_length = get_chrom_offsets(chrom_lengths)
+
+        # Set coverage to False for all positions
+        coverage_array_BQ = np.zeros(genome_length, dtype=bool)
+
+        cmd = [
+        "samtools", "depth",
+        "--threads", str(threads),
+        "-J",
+        "-s",
+        "--min-BQ", str(BQ_threshold), # Only bases with BQ >= threshold count towards depth
+        bam_path
+        ]
+
+        with subprocess.Popen(
+            cmd,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+            text=True
+        ) as proc:
+
+            for line in proc.stdout:
+                chrom, pos_str, depth_str = line.split()
+                pos = int(pos_str) - 1 # Convert position to 0-based
+                depth = int(depth_str)
+                genome_index = offsets[chrom] + pos
+
+                # If depth >= threshold, set coverage to True
+                if depth >= depth_threshold:
+                    coverage_array_BQ[genome_index] = True
+
+        print(f"[INFO] Finished creating BQ coverage array for {bam_path}")
+
+        return coverage_array_BQ
+
+    # Get chromosome lengths from reference FAI
+    chrom_lengths = get_chrom_lengths(ref_fai_path)
+
+    # Get genome length
+    _, genome_length = get_chrom_offsets(chrom_lengths)
+
+    # Create boolean array for include BED coverage
+    include_bed_coverage = coverage_array_bed(include_bed_path, chrom_lengths)
+
+    # Create boolean arrays for EX coverage at given depth and BQ thresholds
+    ex_coverage_depth = coverage_array_depth(ex_dsc_bam_path, chrom_lengths, EX_DEPTH_THRESHOLD, THREADS)
+    ex_coverage_BQ = coverage_array_BQ_threshold(ex_dsc_bam_path, chrom_lengths, EX_DEPTH_THRESHOLD, EX_BQ_THRESHOLD, THREADS)
+
+    # Calculate coverage metrics
+    ex_dsc_coverage_bases = int(np.sum(ex_coverage_depth))
+    ex_dsc_coverage_pct = round(100 * (ex_dsc_coverage_bases / genome_length), ndigits = 2)
+    ex_dsc_high_qual_bases = int(np.sum(ex_coverage_BQ))
+    ex_dsc_high_qual_pct = round(100 * (ex_dsc_high_qual_bases / genome_length), ndigits = 2)
+    ex_dsc_high_qual_unmasked_bases = int(np.sum(ex_coverage_BQ & include_bed_coverage))
+    ex_dsc_high_qual_unmasked_pct = round(100 * (ex_dsc_high_qual_unmasked_bases / genome_length), ndigits = 2)
 
     # Fill JSON data
     json_data = {
-        "sample": sample,
         "total_genome_positions": {
             "description": "Number of positions in the reference genome.",
-            "value": total_genome_positions
+            "value": genome_length
             },
         "ex_dsc_coverage_count": {
             "description": "Number of genome positions with DSC depth > 0.",
@@ -167,7 +231,7 @@ def main(args):
       thickness = 15,
       line = dict(color = "black", width = 0.5),
       label = ["Reference genome positions", "DSC depth > 0", "No DSC depth",
-               f"BQ >= threshold ({BASE_QUALITY_THRESHOLD})", f"BQ < threshold ({BASE_QUALITY_THRESHOLD})", 
+               f"BQ >= threshold ({EX_BQ_THRESHOLD})", f"BQ < threshold ({EX_BQ_THRESHOLD})", 
                "Unmasked", "Masked"],
       color = "blue",
       x = [0.0, 0.4, 0.4, 0.7, 0.7, 1.0, 1.0],
@@ -205,13 +269,14 @@ def main(args):
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument("--bam_ex_dsc", required=True)
+    parser.add_argument("--threads", required=True)
+    parser.add_argument("--ex_dsc_bam", required=True)
     parser.add_argument("--include_bed", required=True)
     parser.add_argument("--ref_fai", required=True)
-    parser.add_argument("--json", required=True)
-    parser.add_argument("--plot", required=True)
-    parser.add_argument("--base_quality_threshold", required=True)
-    parser.add_argument("--sample", required=True)
+    parser.add_argument("--ex_depth_threshold", required=True)
+    parser.add_argument("--ex_bq_threshold", required=True)
+    parser.add_argument("--output_json", required=True)
+    parser.add_argument("--output_plot", required=True)
     parser.add_argument("--log", required=True)
     args = parser.parse_args()
     main(args=args)