+++ CHANGELOG.md
@@ -19,8 +19,8 @@ Types of changes:
 - Added additional unit tests for ms_germline_risk and ms_germline_mask (#280 and 282)
 - Added sort and merge steps to ms_germline_mask to reduce BED size (#283)
 - Added Sankey plot to ex_dsc_coverage_metrics.py (#)
-- Added ex_ms_overlap_metrics.py and test (#)
-- Added coverage_comparisons.py (#)
+- Added ex_coverage_overlap_metrics.py and test (#)
+- Added ex_coverage_by_depth.py and test (#)
 
 ### Changed
 
@@ -33,6 +33,7 @@ Types of changes:
 - Added missing docstrings/comments and standardised style (#281)
 - Removed hardcoded output paths in fastqc_summary_metrics.py, now passed to script by rules (#281)
 - Passed individual parameters to ex_generate_demux_adaptors.py and ex_demux_counts_and_gini.py rather than all of config (#281)
+- Moved overlap and depth metrics from ex_dsc_coverage_metrics.py into new scripts (#)
 - Rewrote ex_dsc_coverage_metrics.py to reduce memory usage (#)
 
 ### Removed